Refactors AllocationCompletionTracker out.
